replace patch-release.go with release.go
we updated the release.go file with the contents from the
patch-release.go file which means we would be using the release.go for
automated releases going forward.
