+++ .github/workflows/patch-release.yml
@@ -19,6 +19,6 @@ jobs:
       - uses: actions/setup-go@v5
         with:
           go-version: ${{ env.GO_VERSION }}
-      - run: make patch-release
+      - run: make release
         env:
           GITHUB_TOKEN: ${{ secrets.GH_TOKEN }}
