+++ Makefile
@@ -199,9 +199,6 @@ crio.conf: bin/crio
 release:
 	${GO_RUN} ./scripts/release
 
-patch-release:
-	${GO_RUN} ./scripts/patch-release
-
 release-notes: ${RELEASE_NOTES}
 	${GO_RUN} ./scripts/release-notes \
 		--output-path ${BUILD_PATH}/release-notes
@@ -552,6 +549,5 @@ metrics-exporter: bin/metrics-exporter
 	bin/metrics-exporter \
 	metrics-exporter \
 	release \
-	patch-release \
 	check-log-lines \
 	verify-dependencies
