+++ scripts/release/release.go
@@ -1,19 +1,19 @@
 package main
 
 import (
+	"bytes"
 	"fmt"
 	"os"
+	"regexp"
 	"strings"
 
 	"github.com/blang/semver/v4"
+	goGit "github.com/go-git/go-git/v5"
 	"github.com/sirupsen/logrus"
 	"sigs.k8s.io/release-sdk/git"
 	"sigs.k8s.io/release-sdk/github"
-	"sigs.k8s.io/release-utils/command"
 	"sigs.k8s.io/release-utils/env"
 	"sigs.k8s.io/release-utils/util"
-
-	"github.com/cri-o/cri-o/internal/version"
 )
 
 const (
@@ -22,7 +22,10 @@ const (
 	orgEnvKey         = "ORG"
 	versionFile       = "internal/version/version.go"
 	branchPrefix      = "release-"
+	crioOrgRepo       = "cri-o"
 )
+
+var releaseMinorVersions = []string{"1.29", "1.28", "1.27"}
 
 func main() {
 	logrus.SetFormatter(&logrus.TextFormatter{DisableTimestamp: true})
@@ -37,46 +40,77 @@ func run() error {
 			"run: $%s environment variable is not set", githubTokenEnvKey,
 		)
 	}
-	if !env.IsSet(orgEnvKey) {
-		return fmt.Errorf(
-			"run: $%s environment variable is not set %s",
-			orgEnvKey,
-			"(should be set to your CRI-O fork organization, like 'gh-name')",
-		)
-	}
+
 	remote := env.Default(gitRemoteEnvKey, "origin")
 	logrus.Infof("Using repository fork remote: %s", remote)
 
-	org := os.Getenv(orgEnvKey)
+	org := env.Default(orgEnvKey, crioOrgRepo)
 	logrus.Infof("Using repository fork organization: %s", org)
+
+	if err := git.ConfigureGlobalDefaultUserAndEmail(); err != nil {
+		return fmt.Errorf("configure global default user and email: %w", err)
+	}
 
 	repo, err := git.OpenRepo(".")
 	if err != nil {
 		return fmt.Errorf("open local repo: %w", err)
 	}
 
-	currentBranch, err := repo.CurrentBranch()
-	if err != nil {
-		return fmt.Errorf("get current branch: %w", err)
-	}
-	logrus.Infof("Using current branch: %s", currentBranch)
+	for _, minorVersion := range releaseMinorVersions {
+		baseBranchName := branchPrefix + minorVersion // returns "release-x.y"
 
-	newVersion, err := incVersion(version.Version, currentBranch)
-	if err != nil {
-		return fmt.Errorf("increment version: %w", err)
-	}
-	logrus.Infof("Using new version: %s", util.SemverToTagString(newVersion))
+		sv, err := getCurrentVersionFromReleaseBranch(repo, baseBranchName) // returns "x.y.z"
+		if err != nil {
+			return fmt.Errorf("current version from release branch %s to semver: %w", minorVersion, err)
+		}
 
-	if err := updateVersionAndCreatePR(
-		repo, newVersion, currentBranch, org, remote,
-	); err != nil {
-		return fmt.Errorf("update version in local repository: %w", err)
+		// Bump up the patch version
+		oldVersion := sv.String()
+		sv.Patch++
+		newVersion := sv.String()
+
+		if err := updateVersionAndCreatePR(
+			repo, newVersion, oldVersion, baseBranchName, org, remote,
+		); err != nil {
+			return fmt.Errorf("update version in local repository: %w", err)
+		}
 	}
 
 	return nil
 }
 
-func incVersion(tag, branch string) (res semver.Version, err error) {
+func getCurrentVersionFromReleaseBranch(repo *git.Repo, baseBranchName string) (res semver.Version, err error) {
+	logrus.Infof("Switching to branch: %s", baseBranchName)
+	if err := repo.Checkout(baseBranchName); err != nil {
+		return res, fmt.Errorf("checkout branch %s: %w", baseBranchName, err)
+	}
+
+	versionFromVersionFile, err := getCurrentVersionFromVersionFile(versionFile) // returns "x.xx.x"
+	if err != nil {
+		return res, fmt.Errorf("reading latest version: %w", err)
+	}
+
+	return convertStringToSemver(versionFromVersionFile)
+}
+
+func getCurrentVersionFromVersionFile(versionFile string) (string, error) {
+	versionPattern := `const\s+Version\s+=\s+"(.+)"`
+
+	content, err := os.ReadFile(versionFile)
+	if err != nil {
+		return "", err
+	}
+
+	re := regexp.MustCompile(versionPattern)
+	matches := re.FindStringSubmatch(string(content))
+	if len(matches) < 2 {
+		return "", fmt.Errorf("unable to find version in %q", versionFile)
+	}
+
+	return matches[1], nil
+}
+
+func convertStringToSemver(tag string) (res semver.Version, err error) {
 	sv, err := util.TagStringToSemver(strings.TrimSpace(tag))
 	if err != nil {
 		return res, fmt.Errorf("convert tag string %s to semver: %w", tag, err)
@@ -84,37 +118,40 @@ func incVersion(tag, branch string) (res semver.Version, err error) {
 
 	// clear any suffix like `-dev`
 	sv.Pre = nil
-
-	if branch == git.DefaultBranch {
-		// New minor version
-		sv.Minor++
-		sv.Patch = 0
-	} else {
-		// New patch version
-		sv.Patch++
-	}
-
 	return sv, nil
 }
 
 func updateVersionAndCreatePR(
-	repo *git.Repo, newVersion semver.Version, branch, org, remote string,
+	repo *git.Repo, newVersion, oldVersion string, baseBranchName, org, remote string,
 ) error {
-	logrus.Info("Updating repository")
+	logrus.Infof("Updating repository from %s to %s", oldVersion, newVersion)
 
-	newBranch := branchPrefix + newVersion.String()
+	newBranch := branchPrefix + newVersion
+	doesTheBranchExistRemotely, err := repo.HasRemoteBranch(newBranch)
+	if err != nil {
+		return fmt.Errorf("remote has branch %s: %w", newBranch, err)
+	}
+
+	if doesTheBranchExistRemotely {
+		// Only Rebase and force push
+		if err := repo.Rebase(newBranch); err != nil {
+			return fmt.Errorf("rebase branch %s: %w", newBranch, err)
+		}
+		opts := &goGit.PushOptions{
+			Force: true,
+		}
+		if err := repo.PushToRemoteWithOptions(opts); err != nil {
+			return fmt.Errorf("force pushing to remote: %s: %w", newBranch, err)
+		}
+		return nil
+	}
 	logrus.Infof("Switching to branch: %s", newBranch)
 	if err := repo.Checkout("-B", newBranch); err != nil {
 		return fmt.Errorf("checkout branch %s: %w", newBranch, err)
 	}
 
 	logrus.Infof("Updating new tag in file %s", versionFile)
-	if err := command.
-		New(
-			"sed", "-i",
-			fmt.Sprintf("s/%s/%s/", version.Version, newVersion),
-			versionFile,
-		).RunSilentSuccess(); err != nil {
+	if err := modifyVersionFile(versionFile, oldVersion, newVersion); err != nil {
 		return fmt.Errorf("update version file: %w", err)
 	}
 
@@ -122,8 +159,9 @@ func updateVersionAndCreatePR(
 	if err := repo.Add(versionFile); err != nil {
 		return fmt.Errorf("add file %s to repo: %w", versionFile, err)
 	}
+
 	if err := repo.UserCommit(
-		"Bump version to " + newVersion.String(),
+		"version: bump to " + newVersion,
 	); err != nil {
 		return fmt.Errorf("commit changes: %w", err)
 	}
@@ -136,18 +174,38 @@ func updateVersionAndCreatePR(
 	logrus.Info("Creating PR")
 	gh := github.New()
 
-	pr, err := gh.CreatePullRequest("cri-o", "cri-o", branch,
-		fmt.Sprintf("%s:%s", org, newBranch),
-		fmt.Sprintf("Bump version for %s", newVersion),
-		fmt.Sprintf(
-			"Automated version bump to version `%s`\n\n%s",
-			newVersion, "/release-note-none",
-		),
+	headBranchName := fmt.Sprintf("%s:%s", org, newBranch)
+	title := "Bump version to " + newVersion
+	body := fmt.Sprintf(
+		"Automated version bump to version `%s`\n\n```release-note\nNone\n```",
+		newVersion,
+	)
+
+	pr, err := gh.CreatePullRequest(org, crioOrgRepo, baseBranchName,
+		headBranchName,
+		title,
+		body,
 	)
 	if err != nil {
 		return fmt.Errorf("create pull request: %w", err)
 	}
 	logrus.Infof("Created PR #%d", pr.GetNumber())
+
+	return nil
+}
+
+func modifyVersionFile(filePath, oldVersion, newVersion string) error {
+	content, err := os.ReadFile(filePath)
+	if err != nil {
+		return err
+	}
+
+	modifiedContent := bytes.ReplaceAll(content, []byte(oldVersion), []byte(newVersion))
+
+	err = os.WriteFile(filePath, modifiedContent, 0o644)
+	if err != nil {
+		return err
+	}
 
 	return nil
 }
